Extract executor module for better separation of concerns
Create dedicated executor module for secure Docker execution:

**New executor module:**
- executor.go (main API - formerly runtime.go)
- volume.go (Docker volume management)
- monitor.go (performance monitoring)
- limitedwriter.go (output limiting)
- init/ (container initialization binary)

**Refactored langs module:**
- Remove api.go (unnecessary wrapper functions)
- Update compile.go to use executor types
- Add executor dependency with local replace

**Benefits:**
- Clear separation: langs=language management, executor=secure execution
- Reusability: executor can be used directly by judge module
- Modularity: each module has focused responsibilities
- Flexibility: judge can control executor directly for advanced use cases

Both modules build successfully with proper dependency management.
