+++ executor/go.mod
@@ -1,0 +1,5 @@
+module github.com/yosupo06/library-checker-judge/executor
+
+go 1.21
+
+require github.com/google/uuid v1.6.0
